+++ src/main/java/christmas/domain/menu/Appetizer.java
@@ -1,9 +1,12 @@
 package christmas.domain.menu;
 
+import christmas.domain.menu.component.Menu;
+import christmas.domain.menu.component.Name;
+import christmas.domain.menu.component.Price;
 import java.util.Arrays;
 import java.util.List;
 
-public enum Appetizer implements Menu{
+public enum Appetizer implements Menu {
     PINE_MUSHROOM_SOUP(new Name("양송이스프"), new Price(6000)),
     TAPAS(new Name("타파스"), new Price(5500)),
     CAESAR_SALAD(new Name("시저샐러드"), new Price(8000));
