+++ src/main/java/christmas/domain/menu/Drink.java
@@ -1,9 +1,12 @@
 package christmas.domain.menu;
 
+import christmas.domain.menu.component.Menu;
+import christmas.domain.menu.component.Name;
+import christmas.domain.menu.component.Price;
 import java.util.Arrays;
 import java.util.List;
 
-public enum Drink implements Menu{
+public enum Drink implements Menu {
     ZERO_COLA(new Name("제로콜라"), new Price(3000), false),
     RED_WINE(new Name("레드와인"), new Price(60000), false),
     CHAMPAGNE(new Name("샴페인"), new Price(25000), true);
