+++ src/main/java/christmas/domain/menu/component/Price.java
@@ -1,4 +1,4 @@
-package christmas.domain.menu;
+package christmas.domain.menu.component;
 
 public class Price {
     private final int value;
